refactor: streamline imports for better organization and consistency
