+++ apps/client/src/app/(dashboards)/videos/page.tsx
@@ -3,14 +3,11 @@
 import { DEFAULT_PAGE } from "@/utils/constants";
 
 import { useGetUserVideos } from "@/features/video/hooks";
-import { usePagination } from "@/hooks/global";
+import { usePagination } from "@/hooks";
 
-import Loader from "@/components/global/loader";
-import PaginationComponent from "@/components/global/pagination";
+import { Loader, PaginationComponent } from "@/components/global";
 
-import AddNoteForm from "@/components/dashboards/add-note-form";
-import Header from "@/components/dashboards/header";
-import NoDataFound from "@/components/dashboards/no-data-found";
+import { AddNoteForm, Header, NoDataFound } from "@/components/dashboards";
 
 import { VideosList } from "@/features/video/components";
